# doc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AriesHub API",
    "version": "0.7.0",
    "description": "Public content catalog, Redis-backed cookie identity, flat paged discussions with comment likes, administrator content and moderation API. Credit balance lives on users and pricing on publications; the credit ledger and unlock tables exist but the credit application API is not implemented yet. Private S3-compatible file uploads and owner-only temporary downloads. All JSON responses use Result {code,msg,data,traceId}; success code is SUCCESS, failures retain semantic HTTP status and business codes. Empty successes use HTTP 200 with data:null; throttling includes Retry-After and data.retryAfterSeconds."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1"
    }
  ],
  "paths": {
...
    "/categories": {
      "get": {
        "operationId": "getCategories",
        "summary": "Categories with published and available publication counts",
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Category"
                      }
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database temporarily unavailable",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/publications": {
      "get": {
        "operationId": "getPublications",
        "summary": "Public publication summaries; never includes body content",
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PublicationPage"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database temporarily unavailable",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10000,
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 24,
              "default": 9
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120,
              "description": "Trimmed literal substring matched against title/summary, publication insensitive. SQL wildcards are escaped."
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]{0,80}$"
            }
          },
          {
            "name": "access",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "",
                "FREE",
                "CREDIT"
              ]
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "",
                "CASE_STUDY",
                "ARTICLE",
                "COURSE"
              ]
            }
          }
        ]
      }
    },
    "/publications/{slug}": {
      "get": {
        "operationId": "getPublication",
        "summary": "Public detail and preview; never includes full body",
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PublicationDetail"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid slug",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing, draft, archived or suspended publication",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database temporarily unavailable",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]{1,120}$"
            }
          }
        ]
      }
    },
    "/publications/{id}/content": {
      "get": {
        "operationId": "getPublicationContent",
        "summary": "Only published and available FREE publications are readable in M1",
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PublicationContent"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Paid content is locked; purchases are not implemented",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing, draft, archived or suspended publication",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database temporarily unavailable",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$",
              "description": "PostgreSQL bigint ID serialized as a string."
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessType": {
        "type": "string",
        "enum": [
          "FREE",
          "CREDIT"
        ]
      },
      "Category": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "slug",
          "name",
          "publicationCount"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9]+$",
            "description": "PostgreSQL bigint ID serialized as a string."
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "publicationCount": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "Preview": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "previewMarkdown",
          "requirements",
          "deliverables",
          "version",
          "updatedAt"
        ],
        "properties": {
          "previewMarkdown": {
            "type": "string"
          },
          "requirements": {
            "type": "string"
          },
          "deliverables": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "msg",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "retryAfterSeconds"
                ],
                "properties": {
                  "retryAfterSeconds": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            ]
          },
          "traceId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "PublicationSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "slug",
          "title",
          "summary",
          "categorySlug",
          "categoryName",
          "publicationType",
          "accessType",
          "creditPrice",
          "publishedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9]+$",
            "description": "PostgreSQL bigint ID serialized as a string."
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "categorySlug": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "accessType": {
            "$ref": "#/components/schemas/AccessType"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publicationType": {
            "$ref": "#/components/schemas/PublicationType"
          },
          "creditPrice": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "PublicationDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "publication",
          "preview"
        ],
        "properties": {
          "preview": {
            "$ref": "#/components/schemas/Preview"
          },
          "publication": {
            "$ref": "#/components/schemas/PublicationSummary"
          }
        }
      },
      "PublicationContent": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "publicationId",
          "markdown",
          "version",
          "updatedAt"
        ],
        "properties": {
          "markdown": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publicationId": {
            "type": "string",
            "pattern": "^[0-9]+$",
            "description": "PostgreSQL bigint ID serialized as a string."
          }
        }
      },
      "PublicationPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "page",
          "size",
          "total",
          "totalPages"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicationSummary"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "size": {
            "type": "integer",
            "minimum": 1,
            "maximum": 24
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "PublicationType": {
        "type": "string",
        "enum": [
          "CASE_STUDY",
          "ARTICLE",
          "COURSE"
        ]
      },
...
    },
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "arieshub_token"
      }
    }
  }
}